requirements.txt
# Requirements for Chainguard Test Project
# Generated from app.py imports

# Cryptography
pycrypto
pycryptodome  # For Crypto module (alternative to pycrypto)

# HTTP/Network
urllib3
requests

# Web Frameworks
Django
Flask

# Authentication/Security
PyJWT  # For jwt module

# Data Processing
numpy
PyYAML  # For yaml module
simplejson
chardet

# Date/Time
pytz
python-dateutil  # For dateutil module

# XML/HTML Parsing
lxml
beautifulsoup4  # For bs4 module

# Image Processing
Pillow  # For PIL module

# Utilities
six
colour-science  # For colour module (alternative: colour)

# Database
SQLAlchemy

# Note: urllib is part of Python standard library, no need to install

